#메인페이지_상세검색_가격범위 항목 추가완료( 관련 property 타입 long -> int 변경완료)

+++ NusangJspProject/src/com/nusang/action/post/Buy_ReservationAction.java
@@ -20,7 +20,7 @@ public class Buy_ReservationAction implements Action {
 		Buy_Reservation br = new Buy_Reservation();
 		User user = ConAsist.getSessionUser(request);
 		int postno = Integer.parseInt(request.getParameter("reser_postno"));
-		long reser_price = Long.parseLong(request.getParameter("reser_price"));
+		int reser_price = Integer.parseInt(request.getParameter("reser_price"));
 		br = Buy_Reservation.builder().user(user).reser_price(reser_price).state(0).build();
 		Integer result = Buy_ReservationDao.getInstance().insertReser(br,postno);
